Fix SRTP double-encryption on send
It turns out that the SRTP send implementation was redundantly encrypting payloads when using the
4-ary `sendPacket` method. The override for that method in `SRTPProtocolImpl` would encrypt the
payload data and then proceed to invoke its parent's implementation. That one in turn would invoke
the 5-ary `sendPacket` method which is also overriden in `SRTPProtocolImpl`. That override would
encrypt the payload a second time, resulting in garbled data on the receiver's end.

Fixed by removing the 4-ary override from `SRTPProtocolImpl`.